refactor: Users 테이블에 name 필드 추가
