+++ src/main/resources/db/migration/V20250111154341__alter_table_users_add_column_name.sql
@@ -1,0 +1,1 @@
+alter table users add column name varchar(10) not null default '홍길동';
